requirements.txt
numpy
umap-learn
sklearn
pandas
matplotlib